Ensure new `tag-branch` workflow triggers downstream actions (#217)
[In GitHub Actions, workflows are not triggered by actions arising from
the default token so an explicit token needs to be provided to achieve
this](https://github.com/orgs/community/discussions/27028).

Without this, when tags are pushed via the actions, no builds are
triggered to rebuild the packages.

This fix and the original change _cannot be tested_ until we next use
them for a release.