Remove `push` build trigger (#231)
When we release, a tag is `push`ed which triggers the build to happen,
using the `tag_image_push` workflow from that release's branch.

When we rebuild a scheduled image, we run `tag_image_push` workflow from
`master`, but supply the specific version as an input - proving that
it's safe to use `tag_image_push` from `master` for all (supported)
versions.

Using a _single_ workflow makes development much easier (e.g. you can
see easily if a secret is used by checking the default branch, rather
than being in _any_ workflow on a maintenance branch).

We can also remove the indirect triggering of the build when
`create-tag` creates a tag to trigger, and instead simply manually
trigger.

Notes:
- will need updating after
https://github.com/hazelcast/management-center-docker/pull/232